Decompose peak detection
Decompose the calculation in 3 separate phases
(data filter - distribution computation - peak detection).

Reorganize source files.
