+++ src/main/scala/userProfiling/UserCallProfiling.scala
@@ -1,4 +1,4 @@
-package pd
+package ta
 
 import com.typesafe.config.{Config, ConfigValueFactory}
 import org.apache.spark._
@@ -70,11 +70,9 @@ object UserCallProfiling {
     val usage = (s"Usage: submit.sh ${appName} <master> <cdrLocation> " +
                  "<geoLocation> <outputLocation> " +
                  s"<baseSince (${Call.datePattern})> " +
-                 s"<baseUntil (${Call.datePattern})> " +
-                 "<maxCores> <driverMem> " +
-                 "<executorMem>")
+                 s"<baseUntil (${Call.datePattern})>")
 
-    if (args.length != 9) {
+    if (args.length != 6) {
       System.err.println(usage)
       System.exit(1)
     }
@@ -85,16 +83,10 @@ object UserCallProfiling {
     val outputLocation=args(3)
     val baseSince = Call.dateFormat.parseDateTime(args(4))
     val baseUntil = Call.dateFormat.parseDateTime(args(5))
-    val maxCores = args(6)
-    val driverMem = args(7)
-    val executorMem = args(8)
 
     val conf = new SparkConf()
       .setAppName(appName)
       .setMaster(master)
-      .set("spark.cores.max", maxCores)
-      .set("spark.driver.memory", driverMem)
-      .set("spark.executor.memory", executorMem)
     val sc = new SparkContext(conf)
 
     val data = sc.textFile(cdrLocation)
